app/requirements.txt
fastapi==0.101.0
FastAPI-SQLAlchemy==0.2.1
langchain==0.0.247
openai==0.27.7
pydantic
python-dotenv
tiktoken==0.4.0
uvicorn==0.23.2